requirements.txt
openai
python-dotenv>=1.2
Pillow>=12.0
opencv-python>=4.0
numpy
PyQt5>=5.15.10
pyttsx3>=2.90
